Added README file  Improved generate_dbd_new.sh script

+++ TESTS/MAIN/hBN/EXCPH_abi/DFT/generate_dbs_new.sh
@@ -1,17 +1,16 @@
+#!/bin/bash
+
 ABINIT="abinit_ibte"
 MRGDDB="mrgddb_ibte"
 MRGDV="mrgdv_ibte"
-A2Y="a2y_a2y"
+A2Y="a2y_excph"
 
 # Control the two possible options from here
 # So far only expand="yes" generates usable GKKP_expanded
 expand="yes" # "no"
 
-
 qkind="q_ibz" # "q_ibz"
-if [ "$expand" == "yes" ] ; then
-  qkind="q_bz"
-fi
+if [ "$expand" == "yes" ] ; then qkind="q_bz" ; fi
 
 if ! test -f io_dir_new/hbn_out_DS1_WFK.nc || ! test -f io_dir_new/hbn_out_DS2_WFK.nc ; then
   rm -rf io_dir_new
@@ -34,23 +33,28 @@ fi
 WFK_file="hbn_out_DS2"
 folder="../../AllSymm"
 if [ "$expand" == "yes" ] ; then
-  echo "expand wfs"
-  $ABINIT < inputs_new/05_wfk_expand.files > 05_wfk_expand.log
   WFK_file="hbn_wfk_expanded_out"
   folder="../../NoSymm"
+  if ! test -f io_dir_new/hbn_wfk_expanded_out_WFK.nc ; then
+    echo "expand wfs"
+    $ABINIT < inputs_new/05_wfk_expand.files > 05_wfk_expand.log
+  fi
 fi
 
 GKKP_file="GKKP"
-if [ "$qkind" == "q_bz" ] ; then
-  GKKP_file="GKKP_expanded"
-fi
-
+if [ "$qkind" == "q_bz" ] ; then GKKP_file="GKKP_expanded" ; fi
 
 echo "generate the yambo SAVE with GKKP_abinit folder"
 cd io_dir_new
-rm hbn_out_DS2_GKKP.nc
-ln -s hbn_${qkind}_out_GSTORE.nc ${WFK_file}_GKKP.nc
+rm -rf ${WFK_file}_GKKP.nc
+if ! test -f ${WFK_file}_GKKP.nc ; then ln -s hbn_${qkind}_out_GSTORE.nc ${WFK_file}_GKKP.nc ; fi
 rm -rf SAVE
 $A2Y -F ${WFK_file}_WFK.nc
-mv SAVE $folder/
-mv GKKP_abinit $folder/$GKKP_file
+if test -d SAVE ; then
+  rm -rf $folder/SAVE
+  rm -rf $folder/02_bse_allq
+  rm -rf $folder/03_excph_lifetimes
+  mv SAVE $folder/
+  rm -rf $folder/$GKKP_file
+  mv GKKP_abinit $folder/$GKKP_file
+fi
